+++ crates/noyalib/examples/rename.rs
@@ -27,7 +27,7 @@ mod snake_case_keys {
 
     pub(super) fn deserialize<'de, T, D>(deserializer: D) -> Result<T, D::Error>
     where
-        T: serde::de::DeserializeOwned,
+        T: serde::de::DeserializeOwned + 'static,
         D: Deserializer<'de>,
     {
         singleton_map_with::deserialize_with(deserializer, singleton_map_with::to_pascal_case)
@@ -65,7 +65,7 @@ mod kebab_case_keys {
 
     pub(super) fn deserialize<'de, T, D>(deserializer: D) -> Result<T, D::Error>
     where
-        T: serde::de::DeserializeOwned,
+        T: serde::de::DeserializeOwned + 'static,
         D: Deserializer<'de>,
     {
         singleton_map_with::deserialize_with(deserializer, singleton_map_with::from_kebab_case)
@@ -104,7 +104,7 @@ mod lowercase_keys {
 
     pub(super) fn deserialize<'de, T, D>(deserializer: D) -> Result<T, D::Error>
     where
-        T: serde::de::DeserializeOwned,
+        T: serde::de::DeserializeOwned + 'static,
         D: Deserializer<'de>,
     {
         singleton_map_with::deserialize_with(deserializer, singleton_map_with::to_uppercase)