+++ crates/noyalib/src/compat/serde_yaml.rs
@@ -219,7 +219,7 @@ pub mod with {
 /// ```
 pub fn from_str<T>(s: &str) -> Result<T>
 where
-    T: DeserializeOwned,
+    T: DeserializeOwned + 'static,
 {
     crate::from_str(s)
 }
@@ -235,7 +235,7 @@ where
 /// ```
 pub fn from_slice<T>(bytes: &[u8]) -> Result<T>
 where
-    T: DeserializeOwned,
+    T: DeserializeOwned + 'static,
 {
     crate::from_slice(bytes)
 }
@@ -258,7 +258,7 @@ where
 pub fn from_reader<R, T>(reader: R) -> Result<T>
 where
     R: std::io::Read,
-    T: DeserializeOwned,
+    T: DeserializeOwned + 'static,
 {
     crate::from_reader(reader)
 }
@@ -280,7 +280,7 @@ where
 /// ```
 pub fn from_value<T>(value: Value) -> Result<T>
 where
-    T: DeserializeOwned,
+    T: DeserializeOwned + 'static,
 {
     crate::from_value(&value)
 }
@@ -362,7 +362,7 @@ where
 /// ```
 pub fn from_str_multi<T>(s: &str) -> Result<Vec<T>>
 where
-    T: DeserializeOwned,
+    T: DeserializeOwned + 'static,
 {
     crate::load_all_as::<T>(s)
 }