+++ crates/noyalib/src/flattened.rs
@@ -135,7 +135,7 @@ impl<T> core::ops::Deref for Flattened<T> {
 
 impl<'de, T> serde::Deserialize<'de> for Flattened<T>
 where
-    T: for<'a> serde::Deserialize<'a>,
+    T: for<'a> serde::Deserialize<'a> + 'static,
 {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where